tasks: feeling-chained, hidden-in-plain-sight-1 && 2 -> translated to x86_64
feeling-chained: keep task on 32-bit for educational purposes

Function arguments are passed on the stack in this case, so the 32-bit
version is pedagogically relevant and kept unchanged.
Updated README.md accordingly.

hidden-in-plain-sight-1 task:
Updated the README.md
Changing the ./support executable file to 64bit

hidden-in-plain-sight-2 task:
Updated the README.md
Changing the ./support executable file to 64bit